requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pymongo==4.5.0
openai==1.3.5
pytest==7.4.0
pytest-flask==1.2.0
python-dateutil==2.8.2
Werkzeug==3.1.3
Jinja2==3.1.2
pyjwt==2.4.0
bson==0.5.10 